IMS_BACKEND/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
django-filter==24.1
drf-yasg==1.21.7
python-dotenv==1.0.1 